Align norminette coding convention
