+++ src/hook/mlx.c
@@ -6,7 +6,7 @@
 /*   By: seunan <seunan@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/10 23:21:30 by sunko             #+#    #+#             */
-/*   Updated: 2023/12/13 22:24:49 by seunan           ###   ########.fr       */
+/*   Updated: 2023/12/13 23:02:50 by seunan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,75 +40,6 @@ static int	destroy_window(t_mlx_data *mlx_data)
 	mlx_destroy_window(mlx_data->mlx, mlx_data->win);
 	exit(EXIT_SUCCESS);
 	return (0);
-}
-
-#include <stdio.h>
-void	move_front(t_mlx_args *mlx_args)
-{
-	t_camera	camera;
-
-	camera = mlx_args->vars->camera;
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_args->vars->camera.view_point = \
-		v_plus(mlx_args->vars->camera.view_point, vec3(0, 0, 0.1));
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_destroy_image(mlx_args->mlx_data->mlx, mlx_args->mlx_data->img);
-	mlx_args->mlx_data->img = mlx_new_image(mlx_args->mlx_data->mlx, WIDTH, HEIGHT);
-	mlx_args->mlx_data->addr = mlx_get_data_addr(mlx_args->mlx_data->img, \
-		&mlx_args->mlx_data->bits_per_pixel, &mlx_args->mlx_data->line_length, &mlx_args->mlx_data->endian);
-	render(mlx_args->vars, mlx_args->mlx_data);
-	mlx_put_image_to_window(mlx_args->mlx_data->mlx, mlx_args->mlx_data->win, mlx_args->mlx_data->img, 0, 0);
-}
-
-void	move_left(t_mlx_args *mlx_args)
-{
-	t_camera	camera;
-
-	camera = mlx_args->vars->camera;
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_args->vars->camera.view_point = \
-		v_plus(mlx_args->vars->camera.view_point, vec3(0.1, 0, 0));
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_destroy_image(mlx_args->mlx_data->mlx, mlx_args->mlx_data->img);
-	mlx_args->mlx_data->img = mlx_new_image(mlx_args->mlx_data->mlx, WIDTH, HEIGHT);
-	mlx_args->mlx_data->addr = mlx_get_data_addr(mlx_args->mlx_data->img, \
-		&mlx_args->mlx_data->bits_per_pixel, &mlx_args->mlx_data->line_length, &mlx_args->mlx_data->endian);
-	render(mlx_args->vars, mlx_args->mlx_data);
-	mlx_put_image_to_window(mlx_args->mlx_data->mlx, mlx_args->mlx_data->win, mlx_args->mlx_data->img, 0, 0);
-}
-
-void	move_back(t_mlx_args *mlx_args)
-{
-	t_camera	camera;
-
-	camera = mlx_args->vars->camera;
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_args->vars->camera.view_point = \
-		v_minus(mlx_args->vars->camera.view_point, vec3(0, 0, 0.1));
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_destroy_image(mlx_args->mlx_data->mlx, mlx_args->mlx_data->img);
-	mlx_args->mlx_data->img = mlx_new_image(mlx_args->mlx_data->mlx, WIDTH, HEIGHT);
-	mlx_args->mlx_data->addr = mlx_get_data_addr(mlx_args->mlx_data->img, \
-		&mlx_args->mlx_data->bits_per_pixel, &mlx_args->mlx_data->line_length, &mlx_args->mlx_data->endian);
-	render(mlx_args->vars, mlx_args->mlx_data);
-	mlx_put_image_to_window(mlx_args->mlx_data->mlx, mlx_args->mlx_data->win, mlx_args->mlx_data->img, 0, 0);
-}
-
-void	move_right(t_mlx_args *mlx_args)
-{
-	t_camera	camera;
-
-	camera = mlx_args->vars->camera;
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_args->vars->camera.view_point = \
-		v_minus(mlx_args->vars->camera.view_point, vec3(0.1, 0, 0));
-	printf("x: %lf, y: %lf, z: %lf\n", camera.view_point.x, camera.view_point.y, camera.view_point.z);
-	mlx_destroy_image(mlx_args->mlx_data->mlx, mlx_args->mlx_data->img);
-	mlx_args->mlx_data->img = mlx_new_image(mlx_args->mlx_data->mlx, WIDTH, HEIGHT);
-	mlx_args->mlx_data->addr = mlx_get_data_addr(mlx_args->mlx_data->img, \
-		&mlx_args->mlx_data->bits_per_pixel, &mlx_args->mlx_data->line_length, &mlx_args->mlx_data->endian);
-	render(mlx_args->vars, mlx_args->mlx_data);
-	mlx_put_image_to_window(mlx_args->mlx_data->mlx, mlx_args->mlx_data->win, mlx_args->mlx_data->img, 0, 0);
 }
 
 static int	handle_key_press(int keycode, t_mlx_args *mlx_args)
